Switch transcoding code to use f64 secs for bounds

+++ src/web_server/services/hls_segment_service.rs
@@ -16,7 +16,7 @@ use std::sync::Arc;
 use std::time::Instant;
 use tracing::info;
 
-pub const SEGMENT_DURATION: i64 = 5;
+pub const SEGMENT_DURATION: f64 = 5.0;
 
 pub const QUALITY_LEVELS: &[HlsQualityLevel] = &[
 	HlsQualityLevel {
@@ -221,7 +221,7 @@ impl ArtifactGenerator for HlsSegmentGenerator {
 			
 			info!("Generating segment {} at {} for {:?}", input.segment_index, input.quality_level.id, &opts.media_path);
 			
-			let start_time = input.segment_index as i64 * SEGMENT_DURATION;
+			let start_time = input.segment_index as f64 * SEGMENT_DURATION;
 			let time_range = start_time..(start_time + SEGMENT_DURATION);
 			
 			transcoding::transcode_segment(opts, time_range)
